execution/backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
sqlalchemy>=2.0.27
psycopg2-binary==2.9.11
asyncpg>=0.29.0
pydantic>=2.6.1
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
redis>=5.0.1
httpx>=0.26.0
python-dotenv>=1.0.1
alembic>=1.13.1
pytest>=8.0.0
aioredis>=2.0.1
